+++ src/main/java/de/focusshift/zeiterfassung/overtime/OvertimeAccountRepository.java
@@ -1,4 +1,4 @@
-package de.focusshift.zeiterfassung.usermanagement;
+package de.focusshift.zeiterfassung.overtime;
 
 import org.springframework.data.jpa.repository.JpaRepository;
